% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_all.R
\name{batch_notebooks}
\alias{batch_notebooks}
\title{Allocate IDs to batches}
\usage{
batch_notebooks(notebooks, batch_size)
}
\arguments{
\item{notebooks}{A character vector of notebook IDs}

\item{batch_size}{Integer with number per batch}
}
\value{
A list of batch vectors
}
\description{
Allocate IDs to batches
}
