% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{build_comments}
\alias{build_comments}
\title{Retrieve comments for a notebook and prepare for solr}
\usage{
build_comments(notebook_info)
}
\arguments{
\item{notebook_info}{A list generated by \code{build_notebook_info} with information from
the notebook that needs to be added to every solr document}
}
\value{
A list with solr ready comment documents
}
\description{
Retrieve comments for a notebook and prepare for solr
}
