% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{build_update_metadata}
\alias{build_update_metadata}
\title{Build the Metadata for a Notebook Update}
\usage{
build_update_metadata(notebook, starcount)
}
\arguments{
\item{notebook}{[notebook] The notebook object as stored by rcloud.support}

\item{starcount}{[numeric] notebook starcount}
}
\value{
A list that conforms with the solr configuration
}
\description{
Build the Metadata for a Notebook Update
}
