% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_all.R
\name{index_all}
\alias{index_all}
\title{Index all RCloud Notebooks}
\usage{
index_all(pause = 1, batch_size = 10, commitWithin = 10000, ...)
}
\arguments{
\item{pause}{time in seconds between updateseturn with the highlighting}

\item{batch_size}{number of documents per update}

\item{commitWithin}{Number of milliseconds before solr forces commit}

\item{...}{Extra arguments passed to \code{.solr.post}}
}
\description{
Sends all notebooks from all users to solr
}
