% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse.solr.res}
\alias{parse.solr.res}
\title{Parse the result from a solr search}
\usage{
parse.solr.res(solr.res, pagesize)
}
\arguments{
\item{solr.res}{The return value from the GET request}

\item{pagesize}{solr pagesize}
}
\value{
The parsed search result
}
\description{
Parse the result from a solr search
}
