% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rcloud.search.description}
\alias{rcloud.search.description}
\title{Passthrough Notebook Search}
\usage{
rcloud.search.description(description, user = NULL, start = 0,
  pagesize = 100, sortby = "description", orderby = "desc")
}
\arguments{
\item{description}{search string to match against description. Fuzzy matching and wildcarding is used.}

\item{user}{optional to specify a user}

\item{start}{Passed to solr}

\item{pagesize}{Passed to solr}

\item{sortby}{Passed to solr for sorting}

\item{orderby}{Passed to solr for sorting}
}
\value{
Search result direct from solr with no parsing
}
\description{
On description and optionally user. This does minimal processing server side to increase speed.
}
\examples{
\dontrun{
rcloud.search.description("test", user = "juninho")
}
}
