% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rcloud.search}
\alias{rcloud.search}
\title{Search RCloud Notebooks}
\usage{
rcloud.search(query, all_sources, sortby, orderby, start, pagesize,
  group.limit = 4)
}
\arguments{
\item{query}{Search string}

\item{all_sources}{Logical: Search single or multiple solr instances?}

\item{sortby}{Passed to solr for sorting}

\item{orderby}{Passed to solr for sorting}

\item{start}{Passed to solr}

\item{pagesize}{Passed to solr}

\item{group.limit}{Passed to solr. Controls how many cells to highlight for each notebook hit.}
}
\value{
Search response after parsing
}
\description{
Main search function exposed as an OCAP to the client.
}
