% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{sc_check_errors}
\alias{sc_check_errors}
\title{Stop and return errors}
\usage{
sc_check_errors(raw_results)
}
\arguments{
\item{raw_results}{Results from solr sources}
}
\description{
Stop and return errors
}
