% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{sc_merge_results}
\alias{sc_merge_results}
\title{Merge Raw Results}
\usage{
sc_merge_results(raw_results, sortby, orderby)
}
\arguments{
\item{raw_results}{A list of results from solr returned from the SearchSource
objects.}

\item{sortby}{Passed to solr for sorting}

\item{orderby}{Passed to solr for sorting}
}
\value{
A single result set, merged and sorted
}
\description{
Take a list of raw results and merge into one big list of results
}
