% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{sc_schema_version}
\alias{sc_schema_version}
\title{Get the schema version}
\usage{
sc_schema_version(path = "schema/version", solr.url, solr.auth.user,
  solr.auth.pwd)
}
\arguments{
\item{path}{The GET route}

\item{solr.url}{Usually from \code{solr.url} config. In testing can be
\code{http://solr:8983/solr/rcloudnotebooks}}

\item{solr.auth.user}{Usually from \code{solr.url} config. \code{NULL}
in testing}

\item{solr.auth.pwd}{Usually from \code{solr.url} config. \code{NULL} in
testing}
}
\value{
The version as a string
}
\description{
Get the schema version
}
