% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{SearchController}
\alias{SearchController}
\title{Search Controller}
\arguments{
\item{main_source}{a list containing at least \code{solr.url}}

\item{gist_sources}{a named list of lists containing at least \code{solr.url}}
}
\description{
A description of this R6 class
}
\section{Details}{

\code{$new(main_source = NULL, gist_sources = NULL)} starts a new search controller

\code{$set_sources(main_source = NULL, gist_sources = NULL)}

\code{$get_sources()} return the config
}

\examples{
# Might need rcloud.support running for this to work
\dontrun{
SC <- SearchController$new(config = list(solr.url="http://solr:8983/solr/rcloudnotebooks"))
}

}
