% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-source.R
\name{SearchSource}
\alias{SearchSource}
\title{Search Source}
\arguments{
\item{src_params}{A list of solr details. Must at least contain
\code{solr.url}, and the source name, \code{source}.}
}
\description{
An R6 class to make requests to individual src_params sources.
}
\section{Details}{

\code{$new(src_name, src_params)} starts a new search source

\code{$search(...)} search this source. Parameters are all based on
  \code{\link{ss_search}}.
}

\examples{
\dontrun{
SS <-
SearchSource$new(
  src_params = list(source =  "main_source",
                    solr.url = "http://solr:8983/solr/rcloudnotebooks")
)
}

}
