% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-source-v1.R
\name{SearchSourceV1}
\alias{SearchSourceV1}
\title{Search a Version 1 Source}
\description{
An R6 class to make requests to individual src_params sources. Child of \code{\link{SearchSource}}
}
\examples{
\dontrun{
SS <-
SearchSourceV1$new(
  src_params = list(source =  "main_source",
                    solr.url = "http://solr:8983/solr/rcloudnotebooks")
)
}

}
