% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-doc.R
\name{solr.delete.comment}
\alias{solr.delete.comment}
\title{Delete Comment from Index}
\usage{
solr.delete.comment(id, cid)
}
\arguments{
\item{id}{Notebook ID}

\item{cid}{Comment ID}
}
\description{
Delete Comment from Index
}
