% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-doc.R
\name{solr.delete.doc}
\alias{solr.delete.doc}
\title{Delete Notebook from Index}
\usage{
solr.delete.doc(id)
}
\arguments{
\item{id}{Notebook ID}
}
\description{
Delete this notebook and all its child docs
}
