% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.solr.get}
\alias{.solr.get}
\title{A GET interface to search solr}
\usage{
.solr.get(query, solr.url = rcloud.support:::getConf("solr.url"),
  solr.auth.user = rcloud.support:::getConf("solr.auth.user"),
  solr.auth.pwd = rcloud.support:::getConf("solr.auth.pwd"))
}
\arguments{
\item{query}{Character. The search term.}

\item{solr.url}{Usually from \code{solr.url} config. In testing can be
\code{http://solr:8983/solr/rcloudnotebooks}}

\item{solr.auth.user}{Usually from \code{solr.url} config. \code{NULL}
in testing}

\item{solr.auth.pwd}{Usually from \code{solr.url} config. \code{NULL} in
testing}
}
\value{
The search response from solr
}
\description{
The standard search method with GET.
}
