% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-doc.R
\name{solr.modify.comment}
\alias{solr.modify.comment}
\title{Modify Index for a Comment}
\usage{
solr.modify.comment(id, content, cid)
}
\arguments{
\item{id}{Notebook ID}

\item{content}{New comment content}

\item{cid}{Comment ID}
}
\description{
Modify Index for a Comment
}
