% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-doc.R
\name{solr.post.comment}
\alias{solr.post.comment}
\title{Index a New Comment}
\usage{
solr.post.comment(id, content, comment.id)
}
\arguments{
\item{id}{Notebook ID}

\item{content}{The new comment content}

\item{comment.id}{The comment ID}
}
\description{
Index a New Comment
}
