% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.solr.post.search}
\alias{.solr.post.search}
\title{A POST interface to search}
\usage{
.solr.post.search(query, solr.url = rcloud.support:::getConf("solr.url"),
  solr.auth.user = rcloud.support:::getConf("solr.auth.user"),
  solr.auth.pwd = rcloud.support:::getConf("solr.auth.pwd"))
}
\arguments{
\item{query}{Character. The search term.}

\item{solr.url}{Usually from \code{solr.url} config. In testing can be \code{http://solr:8983/solr/rcloudnotebooks}}

\item{solr.auth.user}{Usually from \code{solr.url} config. \code{NULL} in testing}

\item{solr.auth.pwd}{Usually from \code{solr.url} config. \code{NULL} in testing}
}
\value{
The search response
}
\description{
Searches are normally made with a GET request. This is a post interface.
}
