% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{solr_core_create}
\alias{solr_core_create}
\title{Create a new solr core}
\usage{
solr_core_create(solr.url = "http://localhost:8983",
  name = "rcloudnotebooks", overwrite = FALSE,
  solr_dir = "/data/rcloud/solrdata", data_dir = "data",
  conf_dir = "conf")
}
\arguments{
\item{solr.url}{base URL of the solr instance (e.g. "http://localhost:8983")}

\item{name}{name of the notebook core. This should probably be left as
"rcloudnotebooks" unless you know what you're doing.}

\item{overwrite}{Logical. If \code{FALSE} then will stop if the core already
exists. If \code{TRUE} this will only overwrite the conf directory. Data
will be left intact.}

\item{solr_dir}{Full path to the solr directory. This doesn't need to be the
same as the directory where the binaries are stored.}

\item{data_dir}{Relative path. Name of the directory, within the core, where
the data will be stored. No real need to change this.}

\item{conf_dir}{Relative path. Name of the directory, within the core, where
the configuration files will be stored. No real need to change this.}
}
\description{
This needs to be run on the system that runs solr. It will create a new data
 directory, copy the configuration in from the \code{inst/solr} package
 directory and make a call to the \code{CREATE} solr API.
}
