% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-source.R
\name{ss_join_docs_high}
\alias{ss_join_docs_high}
\title{Join the grouped documents to the highlighted documents}
\usage{
ss_join_docs_high(docs, highlight)
}
\arguments{
\item{docs}{The groups object from the solr response object}

\item{highlight}{The highlighting object from the solr response object}
}
\value{
A list similar to \code{docs} but with highlighting added in
}
\description{
Join the grouped documents to the highlighted documents
}
