% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ss_search}
\alias{ss_search}
\title{Search a single source}
\usage{
ss_search(query, solr.url, solr.auth.user = NULL, solr.auth.pwd = NULL,
  sortby, orderby, start = 0, pagesize = 10, group.limit = 4,
  hl.fragsize = 60)
}
\arguments{
\item{query}{Search string}

\item{solr.url}{URL for the solr request}

\item{solr.auth.user}{Solr Authentication, username}

\item{solr.auth.pwd}{Solr Authentication, password}

\item{sortby}{Passed to solr for sorting}

\item{orderby}{Passed to solr for sorting}

\item{start}{Passed to solr}

\item{pagesize}{Passed to solr}

\item{group.limit}{Passed to solr. Controls how many cells to highlight for each notebook hit.}

\item{hl.fragsize}{How many charachters to return with the highlighting}
}
\value{
Search response after parsing
}
\description{
For a single source, query solr and return results. For a single source system
this should be the same as \code{rcloud.search}
}
