% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{update.solr}
\alias{update.solr}
\title{Update the index after a notebook change}
\usage{
\method{update}{solr}(notebook, starcount)
}
\arguments{
\item{notebook}{The notebook object as stored by rcloud.support}

\item{starcount}{notebook starcount}
}
\description{
Update the index after a notebook change
}
