% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_batch}
\alias{update_batch}
\title{Update index for a batch of notebooks}
\usage{
update_batch(batch, detach = FALSE, ...)
}
\arguments{
\item{batch}{A character vector of notebook IDs}

\item{detach}{[logical] Should parallel process be detached? Use FALSE for testing or batch.}

\item{...}{Extra arguments passed on to \code{\link{.solr.post}}.}
}
\value{
response from .solr.post
}
\description{
Update index for a batch of notebooks
}
