% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_solr}
\alias{update_solr}
\title{Update the index after a notebook change}
\usage{
update_solr(notebook, starcount, detach = TRUE, ...)
}
\arguments{
\item{notebook}{[notebook] The notebook object as stored by rcloud.support}

\item{starcount}{[numeric] notebook starcount}

\item{detach}{[logical] Should parallel process be detached? Use FALSE for testing or batch.}

\item{...}{Extra arguments passed on to \code{\link{.solr.post}}.}
}
\description{
Update the index after a notebook change
}
