% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_all.R
\name{index_all}
\alias{index_all}
\title{Index all RCloud Notebooks}
\usage{
index_all(pause = 0.5)
}
\arguments{
\item{pause}{time in seconds between updateseturn with the highlighting}
}
\value{
Search response after parsing
}
\description{
Sends all notebooks from all users to solr
}
