% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{sc_search}
\alias{sc_search}
\title{Main search interface}
\usage{
sc_search(self, private, all_sources, start, pagesize, sortby, orderby, ...)
}
\arguments{
\item{self}{pointer to this object}

\item{private}{private members}

\item{all_sources}{Logical: Search single or multiple solr instances?}

\item{start}{Passed to solr}

\item{pagesize}{Passed to solr}

\item{sortby}{Passed to solr for sorting}

\item{orderby}{Passed to solr for sorting}

\item{...}{arguments to pass to \code{ss_search}}
}
\description{
Called by rcloud.search
}
