% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{SearchController}
\alias{SearchController}
\title{Search Controller}
\description{
A description of this R6 class
}
\section{Details}{

\code{$new()} starts a new search controller

\code{$set_config()} retrieves settings from rcloud.support or allows them to be specified by the user

\code{$set_config()} return the config
}

\examples{
# Might need rcloud.support running for this to work
\dontrun{
SC <- SearchController$new(config = list(solr.url="http://solr:8983/solr/rcloudnotebooks"))
}

}
