% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{SearchController}
\alias{SearchController}
\title{Search Controller}
\arguments{
\item{sources}{a named list of sources. Each source must be a list containing
at least \code{solr.url}. The first source must be called \code{main_source}.}
}
\description{
An R6 class to control the gathering, sorting and paging of search results.
}
\section{Details}{

\code{$new(sources = NULL)} starts a new search controller

\code{$set_sources(sources)}
 If sources is provided they are concatenated they are
 retrieved from \code{rcloud.config} (\code{main_source})
 and \code{.session} (\code{gist_sources}).

\code{$get_sources()} return the config
}

\examples{
\dontrun{
SC <-
SearchController$new(sources = list(
 main_source = list(solr.url = "http://solr:8983/solr/rcloudnotebooks")
))
}

}
