% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-controller.R
\name{SearchController}
\alias{SearchController}
\title{Search Controller}
\description{
A description of this R6 class
}
\section{Details}{

\code{$new()} starts a new search controller
}

\examples{
# Might need rcloud.support running for this to work
\dontrun{
SC <- SearchController$new(something)
}

}
