% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-source.R
\name{SearchSource}
\alias{SearchSource}
\title{Search Source}
\description{
An R6 class to make requests to individual solr sources.
}
\section{Details}{

\code{$new(source = NULL)} starts a new search source
}

\examples{
\dontrun{
SS <- SearchSource$new(source list(solr.url = "http://solr:8983/solr/rcloudnotebooks"))
}

}
